requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20240314
python-dotenv>=1.0.0
requests>=2.31.0
pydantic-ai>=0.0.14
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
anthropic>=0.40.0
gradio>=4.0.0
pytest>=8.0.0
